Move lookup/definition processors, helpers, and build items into .component subpackage
Group the component-model classes (lookup, definition, request processors
and their supporting build items and helpers) into a dedicated .component
subpackage in each extension's deployment and deployment-spi modules.
