+++ extensions/datasource/deployment/src/main/java/io/quarkus/datasource/deployment/component/DataSourceLookupProcessor.java
@@ -1,11 +1,11 @@
-package io.quarkus.datasource.deployment;
+package io.quarkus.datasource.deployment.component;
 
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
 
-import io.quarkus.datasource.deployment.spi.DataSourceLookupBuildItem;
-import io.quarkus.datasource.deployment.spi.DataSourceRequestHandlerBuildItem;
+import io.quarkus.datasource.deployment.spi.component.DataSourceLookupBuildItem;
+import io.quarkus.datasource.deployment.spi.component.DataSourceRequestHandlerBuildItem;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.component.ComponentLookup;
 import io.quarkus.runtime.util.ProgrammingParadigm;
@@ -17,8 +17,8 @@ import io.quarkus.runtime.util.Reason;
  * {@link DataSourceRequestHandlerBuildItem} contributions provided by extension-specific
  * processors (Agroal for JDBC, reactive-datasource for reactive).
  *
- * @see io.quarkus.agroal.deployment.DataSourceDefinitionBlockingProcessor
- * @see io.quarkus.reactive.datasource.deployment.DataSourceDefinitionReactiveProcessor
+ * @see io.quarkus.agroal.deployment.component.DataSourceDefinitionBlockingProcessor
+ * @see io.quarkus.reactive.datasource.deployment.component.DataSourceDefinitionReactiveProcessor
  */
 class DataSourceLookupProcessor {
