+++ extensions/hibernate-orm/deployment-spi/src/main/java/io/quarkus/hibernate/orm/deployment/spi/component/PersistenceUnitRequestBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.deployment.spi;
+package io.quarkus.hibernate.orm.deployment.spi.component;
 
 import io.quarkus.builder.item.MultiBuildItem;
 import io.quarkus.runtime.util.ProgrammingParadigm;
